speech_and_text/scripts/[TTS] Client-Service/Larynx/Larynx-TTS-Docker/requirements.txt
terminal-colored-print
websocket-client
